steps/06_forward_fill_udf/requirements.snowflake.txt
snowflake-snowpark-python
snowflake
python-dotenv